[#226] Do not register ExchangeObserver on unsent requests.
Made sure that the Matchers register their ExchangeObservers only on
outbound Exchanges that have been registered with the
MessageExchangeStore successfully.

When canceling a request that has not been registered successfully, e.g.
because there are no MIDs left to be used, there is no ExchangeObserver
to be invoked at all.